move and rename CurrentDamFolderContext
